some minor refactors and rewording of docstrings.
